+++ bigframes/core/groupby/__init__.py
@@ -195,7 +195,7 @@ class DataFrameGroupBy(vendored_pandas_groupby.DataFrameGroupBy):
 
     def shift(self, periods=1) -> series.Series:
         window = core.WindowSpec(
-            grouping_keys=self._by_col_ids,
+            grouping_keys=tuple(self._by_col_ids),
             preceding=periods if periods > 0 else None,
             following=-periods if periods < 0 else None,
         )
@@ -203,7 +203,7 @@ class DataFrameGroupBy(vendored_pandas_groupby.DataFrameGroupBy):
 
     def diff(self, periods=1) -> series.Series:
         window = core.WindowSpec(
-            grouping_keys=self._by_col_ids,
+            grouping_keys=tuple(self._by_col_ids),
             preceding=periods if periods > 0 else None,
             following=-periods if periods < 0 else None,
         )
@@ -212,7 +212,7 @@ class DataFrameGroupBy(vendored_pandas_groupby.DataFrameGroupBy):
     def rolling(self, window: int, min_periods=None) -> windows.Window:
         # To get n size window, need current row and n-1 preceding rows.
         window_spec = core.WindowSpec(
-            grouping_keys=self._by_col_ids,
+            grouping_keys=tuple(self._by_col_ids),
             preceding=window - 1,
             following=0,
             min_periods=min_periods or window,
@@ -227,7 +227,7 @@ class DataFrameGroupBy(vendored_pandas_groupby.DataFrameGroupBy):
 
     def expanding(self, min_periods: int = 1) -> windows.Window:
         window_spec = core.WindowSpec(
-            grouping_keys=self._by_col_ids,
+            grouping_keys=tuple(self._by_col_ids),
             following=0,
             min_periods=min_periods,
         )
@@ -391,7 +391,7 @@ class DataFrameGroupBy(vendored_pandas_groupby.DataFrameGroupBy):
     ):
         """Apply window op to groupby. Defaults to grouped cumulative window."""
         window_spec = window or core.WindowSpec(
-            grouping_keys=self._by_col_ids, following=0
+            grouping_keys=tuple(self._by_col_ids), following=0
         )
         columns = self._aggregated_columns(numeric_only=numeric_only)
         block, result_ids = self._block.multi_apply_window_op(
@@ -531,7 +531,7 @@ class SeriesGroupBy(vendored_pandas_groupby.SeriesGroupBy):
     def shift(self, periods=1) -> series.Series:
         """Shift index by desired number of periods."""
         window = core.WindowSpec(
-            grouping_keys=self._by_col_ids,
+            grouping_keys=tuple(self._by_col_ids),
             preceding=periods if periods > 0 else None,
             following=-periods if periods < 0 else None,
         )
@@ -539,7 +539,7 @@ class SeriesGroupBy(vendored_pandas_groupby.SeriesGroupBy):
 
     def diff(self, periods=1) -> series.Series:
         window = core.WindowSpec(
-            grouping_keys=self._by_col_ids,
+            grouping_keys=tuple(self._by_col_ids),
             preceding=periods if periods > 0 else None,
             following=-periods if periods < 0 else None,
         )
@@ -548,7 +548,7 @@ class SeriesGroupBy(vendored_pandas_groupby.SeriesGroupBy):
     def rolling(self, window: int, min_periods=None) -> windows.Window:
         # To get n size window, need current row and n-1 preceding rows.
         window_spec = core.WindowSpec(
-            grouping_keys=self._by_col_ids,
+            grouping_keys=tuple(self._by_col_ids),
             preceding=window - 1,
             following=0,
             min_periods=min_periods or window,
@@ -567,7 +567,7 @@ class SeriesGroupBy(vendored_pandas_groupby.SeriesGroupBy):
 
     def expanding(self, min_periods: int = 1) -> windows.Window:
         window_spec = core.WindowSpec(
-            grouping_keys=self._by_col_ids,
+            grouping_keys=tuple(self._by_col_ids),
             following=0,
             min_periods=min_periods,
         )
@@ -600,7 +600,7 @@ class SeriesGroupBy(vendored_pandas_groupby.SeriesGroupBy):
     ):
         """Apply window op to groupby. Defaults to grouped cumulative window."""
         window_spec = window or core.WindowSpec(
-            grouping_keys=self._by_col_ids, following=0
+            grouping_keys=tuple(self._by_col_ids), following=0
         )
 
         label = self._value_name if not discard_name else None
